+++ CHANGES.rst
@@ -17,6 +17,7 @@ development
 - Remove redundant "designated_states" field by using dotted name resolver
   for accessing "register.designated_states" when generating worldmap data
 - Add support for PivotTable data generation in ad-hoc mode with OPSExchangeDocument data model
+- Refactor maps.py and tables.py to "p2n.formatter" namespace
 
 
 2017-12-01 3.0.0-dev5